refactoring template compiler to separate out node-isms
