+++ packages/core/src/browser-index.ts
@@ -2,3 +2,4 @@ export type { default as Package } from './package';
 export { getOrCreate } from './get-or-create';
 export { explicitRelative, extensionsPattern } from './paths';
 export { default as PackageCache } from './browser-package-cache';
+export { TemplateCompiler } from './template-compiler';
