+++ packages/core/src/index.ts
@@ -9,12 +9,8 @@ export { Resolver } from './resolver';
 export { AppMeta, AddonMeta } from './metadata';
 export { default as Package, V2AddonPackage as AddonPackage, V2AppPackage as AppPackage, V2Package } from './package';
 export { default as Stage } from './stage';
-export {
-  TemplateCompiler,
-  Plugins as TemplateCompilerPlugins,
-  TemplateCompilerParams,
-  templateCompilerModule,
-} from './template-compiler';
+export { TemplateCompiler, TemplateCompilerParams, templateCompilerModule } from './template-compiler';
+export { Plugins as TemplateCompilerPlugins } from './ember-template-compiler-types';
 export { Asset, EmberAsset, ImplicitAssetPaths } from './asset';
 export { default as Options, optionsWithDefaults } from './options';
 export { default as toBroccoliPlugin } from './to-broccoli-plugin';
